test siphash24 and siphash13 properly
includes testing against the SipHash 2-4 reference test vector,
see https://131002.net/siphash/siphash24.c

Note this patch includes a change to how we convert a 64 bit SipHash
hash to a 32 bit hash as used by perl. Previously we used the low
bytes only, and with this patch we use an xor of the high and low
bytes. This patch ALSO splits the logic for the 32 bit and 64 bit
variants into different functions, so that we can test against the
reference test vector properly.